VideoRecordingContext ddded in order to manage pointers display
